// apps/mcp-server/src/types.ts
export type ToolResponse = {
  /** 响应内容 */
  content: {
    /** 内容类型 */
    type: 'text';
    /** 文本内容 */
    text: string;
  }[];
};